Fixing async data load issue
Note that this code is still asynchronous (if the forEach loop took very long to run, the choropleth wouldn't render properly). However, since this commit replaces the (slow) nested for loops with more efficient for loop + filter, async is no longer an issue, it seems. Let me know if this doesn't work, or doesn't make sense.